mbllen_service/requirements_realesrgan_alt.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
pillow
torch>=1.7
torchvision
realesrgan
basicsr
facexlib
gfpgan 